form_fields: remove redundant Flexible in Rows
